Change node position to separate key-value pairs
I made a minor change in the web-app so that when posting to Python the node position would be returned as x and y coordinates as opposed to a `pos` vector. This pull request just makes sure that the network can still be plotted with matplotlib after this change has been made.

The purpose is that the output format then resembles more the input one would provide to netwulf.visualize if one wanted to have fixed node positions, something I want to implement soon. So with this change, one can decide on a layout one likes, save the node-link dictionary and easily put it back into netwulf, maybe with different node colors or link structure.